requirements.txt
ffmpeg-python
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
zipfile36
opencv-python
matplotlib
albumentations
torch
torchvision
pytorchvideo
pytorch-lightning
pytorch_lightning
requests
openai-whisper
pandas
yacs>=0.1.6
pyyaml>=5.1
av
matplotlib
termcolor>=1.1
simplejson
tqdm
psutil
matplotlib
opencv-python
pandas
pillow
tensorboard
fairscale
moviepy
git+https://github.com/facebookresearch/pytorchvideo.git
git+https://github.com/facebookresearch/detectron2.git
rich
choix==0.3.5
numpy==1.26.3
openai==1.10.0
scipy==1.12.0
ollama
torch
flask
seaborn
pathlib
torchvision
gunicorn
Jinja2
MarkupSafe
anthropic